Remove tasks test added
