+++ src/functionality.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-use-before-define */
 import {
-  updateTask, handleCheckbox, focus, blur, removeTasks,
+  updateTask, handleCheckbox, focus, blur,
 } from './util.js';
 
 export const addTaskDom = (description, index, completed) => {
@@ -66,6 +66,22 @@ export const createTasksList = () => {
       addTaskDom(task.description, task.index, task.completed);
     });
   }
+};
+
+export const removeTasks = (tasks) => {
+  const tasksListEl = document.getElementById('tasks-list')
+  let tasksList = JSON.parse(localStorage.getItem('tasks'));
+  tasks.forEach(task => {
+    tasksListEl.removeChild(task);
+    const index = parseInt(task.id.slice(-1), 10);
+    tasksList = tasksList.filter((item) => item.index !== index);
+  });
+  
+  const newTasksList = tasksList.map((task, i) => {
+    task.index = i + 1;
+    return task;
+  });
+  localStorage.setItem('tasks', JSON.stringify(newTasksList));
 };
 
 export const removeCompletedTasks = () => {
